[haskell] Removed inputsFrom, which improves the flake usability
The inputsFrom part of the dev shell has been removed. This caused the
package to need to be built for the dev shell to be used, which we
don't want. The up shot of this is now it is much faster to load the
flake environment and the package won't be built after every change.